pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.backends.legacy:build"

[project]
name = "deskbot"
version = "0.1.0"
description = "DeskBot — OpenEnv dual-arm robot desk cleaning environment"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Rohit Suthar" }]
requires-python = ">=3.10"
keywords = ["openenv", "reinforcement-learning", "robotics", "pybullet"]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "fastapi>=0.111.0",
    "uvicorn[standard]>=0.30.1",
    "pydantic>=2.7.1",
    "pyyaml>=6.0.1",
    "websockets>=12.0",
]

[project.optional-dependencies]
simulation = [
    "pybullet>=3.2.6",
    "numpy>=1.26.4",
    "robot_descriptions>=1.9.0",
]
training = [
    "openai>=1.30.0",
    "torch>=2.3.0",
    "stable-baselines3>=2.3.2",
    "gymnasium>=0.29.1",
]
dev = [
    "pytest>=8.2.0",
    "httpx>=0.27.0",
]

[project.urls]
Homepage = "https://huggingface.co/spaces/rohitsuthar/deskbot"
Repository = "https://huggingface.co/spaces/rohitsuthar/deskbot"

[tool.setuptools.packages.find]
where = ["."]
include = ["deskbot*"]
